requirements.txt
python-dotenv==1.0.1
py-cord[voice]==2.6.1
langdetect==1.0.9
textblob==0.17.1
spacy==3.7.2
openai==1.30.1